simple-strawberry-service/pyproject.toml
[project]
name = "graphql-counter"
version = "0.1.0"
description = "Illustrates subscription issue"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "strawberry-graphql (>=0.278.0,<0.279.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "fastapi (>=0.116.1,<0.117.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
